+++ lib/pleroma/web/activity_pub/activity_pub.ex
@@ -199,6 +199,16 @@ defmodule Pleroma.Web.ActivityPub.ActivityPub do
             Pleroma.Web.Streamer.stream("public:local:media", activity)
           end
         end
+<<<<<<< HEAD
+=======
+      else
+        if !Enum.member?(activity.data["cc"] || [], public) &&
+             !Enum.member?(
+               activity.data["to"],
+               User.get_cached_by_ap_id(activity.data["actor"]).follower_address
+             ),
+           do: Pleroma.Web.Streamer.stream("direct", activity)
+>>>>>>> develop
       end
     else
       if !Enum.member?(activity.data["cc"] || [], public) &&
@@ -991,7 +1001,7 @@ defmodule Pleroma.Web.ActivityPub.ActivityPub do
   end
 
   def make_user_from_ap_id(ap_id) do
-    if _user = User.get_by_ap_id(ap_id) do
+    if _user = User.get_cached_by_ap_id(ap_id) do
       Transmogrifier.upgrade_user_from_ap_id(ap_id)
     else
       with {:ok, data} <- fetch_and_prepare_user_from_ap_id(ap_id) do